Added scope related tests
ThirdPartyRulesTest.java:
- combined all third-party rules inside ThirdPartyRulesTest.java
- deleted distinct rule test files
- changed all ArchRules to public as a part of the preparation for task #19
- rules 1, 3, and 4 utilize now the present PackageStructure type.
--- rules 1 and 4 also test scope-related conditions for classic architecture projects now.
